requirements.txt
alpaca-py>=0.30.0
pandas>=2.0.0
python-dotenv>=1.0.0
numpy>=1.26.0
pytz>=2024.1

# dashboard
streamlit>=1.35.0
plotly>=5.22.0

# dev / testing
pytest>=8.0.0
